cpp_yandex red belt cource project

+++ cpp_yandex/courses/3_red_belt/week6/course_project/search_server.h
@@ -7,12 +7,14 @@
 #include <vector>
 #include <map>
 #include <string>
+#include <numeric>
+
 using namespace std;
 
 class InvertedIndex {
 public:
   void Add(string& document);
-  const map<size_t, size_t>& Lookup(const string& word, const map<size_t, size_t>& res) const;
+  const map<int, size_t>& Lookup(const string& word, const map<int, size_t>& res) const;
 
   const string& GetDocument(size_t id) const {
     return docs[id];
@@ -21,7 +23,7 @@ public:
   size_t doc_size() { return docs.size(); }
 
 private:
-  map<string, map<size_t, size_t>> index;
+  map<string, map<int, size_t>> index;
   vector<string> docs;
 };
